% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mutation_types.R
\name{all_snv_mutation_types}
\alias{all_snv_mutation_types}
\title{Gets a list of the standard 96 mutation types used in SNV mutation signature analysis.}
\usage{
all_snv_mutation_types()
}
\value{
A 96-element character vector containing SNV mutation types
}
\description{
Gets a list of the standard 96 mutation types used in SNV mutation signature analysis.
}
