% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mutation_types.R
\name{base_complement}
\alias{base_complement}
\title{Returns the complementary sequence for a set of strings}
\usage{
base_complement(base_strings, reverse = FALSE)
}
\arguments{
\item{base_strings}{Character vector of sequences to return the complement of.
Any characters aside from A, C, G, T will be returned as is.}

\item{reverse}{If TRUE, then the sequence is reversed before it is returned.}
}
\value{
Character vector of complement or reverse complement sequences
}
\description{
Returns the complementary sequence for a set of strings
}
