% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_signatures_by_bleed.R
\name{collapse_signatures_by_bleed}
\alias{collapse_signatures_by_bleed}
\title{Collapse signatures using bleed information}
\usage{
collapse_signatures_by_bleed(exposure_mcmc_output, bleed_threshold = 0.4)
}
\arguments{
\item{exposure_mcmc_output}{Output from get_exposures}

\item{bleed_threshold}{The higher this is set, the fewer levels of collapsing will be attempted}
}
\value{
New exposure object corresponding to the collapsed solution with
        greatest WAIC
}
\description{
Collapse signatures using bleed information
}
