% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_signatures_by_bleed.R
\name{collapse_signatures_by_bleed_all_solutions}
\alias{collapse_signatures_by_bleed_all_solutions}
\title{Collapse signatures using bleed information}
\usage{
collapse_signatures_by_bleed_all_solutions(exposure_mcmc_output,
  bleed_threshold = 0.4)
}
\arguments{
\item{exposure_mcmc_output}{Output from get_exposures}

\item{bleed_threshold}{The higher this is set, the fewer levels of collapsing will be attempted}
}
\value{
List with all attempted collapsed exposure objects
}
\description{
This is a convenience function provided. It will return all levels of bleed
collapsing attempted as a single large list. For a more convenient
implementation, use collapse_signatures_by_bleed, which auto-selects the
optimal solution using the WAIC statistic.
}
