% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_populations.R
\name{compute_population_bic}
\alias{compute_population_bic}
\title{Bayesian Information Criterion for Population Models}
\usage{
compute_population_bic(mutation_table, population_mcmc_output)
}
\arguments{
\item{mutation_table}{Table of mutation data, same as input for 
\code{\link{get_population_signatures}}.}

\item{population_mcmc_output}{Output from \code{\link{get_populations}}.}
}
\value{
BIC value
}
\description{
Bayesian Information Criterion for Population Models
}
