% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_populations.R
\name{compute_population_model_log_likelihood}
\alias{compute_population_model_log_likelihood}
\title{Log Likelihood of Population Model}
\usage{
compute_population_model_log_likelihood(mutation_table, population_mcmc_output)
}
\arguments{
\item{mutation_table}{Table of mutation data, same as input for 
\code{\link{get_population_signatures}}.}

\item{population_mcmc_output}{Output from \code{\link{get_populations}}.}
}
\value{
Log likelihood value associated with data, model, and parameter estimates
}
\description{
Reports the log likelihood of data based on population model parameter estimates
}
