% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_population_signatures.R
\name{compute_population_signatures_bic}
\alias{compute_population_signatures_bic}
\title{Bayesian Information Criterion for Population Signatures}
\usage{
compute_population_signatures_bic(mcmc_output)
}
\arguments{
\item{population_mcmc_output}{Output from \code{\link{get_population_signatures}}.}
}
\value{
BIC value
}
\description{
Bayesian Information Criterion for Population Signatures
}
