% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_population_signatures.R
\name{compute_population_signatures_log_likelihood}
\alias{compute_population_signatures_log_likelihood}
\title{Log Likelihood of Population Signatures Model}
\usage{
compute_population_signatures_log_likelihood(mcmc_output)
}
\arguments{
\item{mcmc_output}{Output from \code{\link{get_population_signatures}}.}
}
\value{
Log likelihood value associated with data, model, and parameter estimates
}
\description{
Reports the log likelihood of data based on population model parameter estimates
}
