% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_population_signatures.R
\name{compute_population_signatures_waic}
\alias{compute_population_signatures_waic}
\title{Watanabe-Akaike Information Criterion for Population Signatures}
\usage{
compute_population_signatures_waic(mcmc_output, parallel = FALSE,
  n_cores = 1)
}
\arguments{
\item{population_mcmc_output}{Output from \code{\link{get_population_signatures}}.}
}
\value{
WAIC value
}
\description{
Watanabe-Akaike Information Criterion for Population Signatures
}
