% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exposures.R
\name{compute_signatures_waic}
\alias{compute_signatures_waic}
\alias{compute_signatures_waic}
\title{Watanabe-Akaike Information Criterion for Signatures}
\usage{
compute_signatures_waic(mcmc_output)

compute_signatures_waic(mcmc_output)
}
\arguments{
\item{mcmc_output}{Output from \code{\link{get_exposures}}.}

\item{mcmc_output}{Output from \code{\link{get_exposures}}.}
}
\value{
WAIC value

WAIC value
}
\description{
Watanabe-Akaike Information Criterion for Signatures

Watanabe-Akaike Information Criterion for Signatures
}
