% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_clustering.R
\name{density_clustering}
\alias{density_clustering}
\title{Cluster MCMC solutions by density}
\usage{
density_clustering(exposures_mcmc_output, minPts = 50,
  feature_selection_threshold = 0.05)
}
\arguments{
\item{exposures_mcmc_output}{Output from get_exposures()}

\item{minPts}{Minimum number of points in a neighbourhood
for inclusion in a density cluster. The default
is 40. Increasing this value will result in
more clusters and more outliers.}

\item{feature_selection_threshold}{For each signature, the percentage of total mutation
burden is computed across all MCMC iterations. If the
maximum exposure is less than the feature selection
threshold, then the signature is removed from density
clustering.}
}
\value{
The HDBSCAN output, which contains data about the clusters corresponding to MCMC exposures
}
\description{
Uses the HDBSCAN method to perform density clustering. This analysis can be
used on the MCMC solution space to identify density clusters which may
represent alternative solutions of interest.
}
