% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster_best_fits.R
\name{get_cluster_best_fits}
\alias{get_cluster_best_fits}
\title{Report the point in each cluster with the lowest residual sum of squares}
\usage{
get_cluster_best_fits(exposures_mcmc_output, exposure_mcmc_clusters)
}
\arguments{
\item{exposures_mcmc_output}{Output from get_exposures() function.}

\item{exposure_mcmc_clusters}{Output from density_clustering() function.}
}
\value{
A dataframe of solutions, one per cluster, in long format (one row per signature per cluster)
}
\description{
Given the known mutation catalog (m), a reference signature matrix (S) and
a potential exposure vector (e), the residual sum of squares (RSS) is defined as
}
\details{
\deqn{\lVert \mathbf{m} - \mathbf{Se} \rVert^2}

such that minimizing RSS produces the "best fit" in terms of Euclidean distance
between between the actual mutation catalog and the catalog simulated using the
computed exposure matrix.
}
