% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster_max_densities.R
\name{get_cluster_max_densities}
\alias{get_cluster_max_densities}
\title{Reports the maximum density point in each cluster}
\usage{
get_cluster_max_densities(exposures_mcmc_output, exposure_mcmc_clusters,
  k_value = 400)
}
\arguments{
\item{exposures_mcmc_output}{Output from plot_exposures()}

\item{exposure_mcmc_clusters}{Output from density_clustering()}

\item{k_value}{The value of k (default 400)}
}
\value{
A dataframe of solutions, one per cluster, in long format (one row per signature per cluster)
}
\description{
As MCMC chains converge, sampling density becomes proportional to the posterior distribution.
Thus, areas of greater density in the MCMC solution space may represent more promising solutions.
As a simple heuristic for density, this function returns the point with the least mean distance
to its nearest k neighbours. By default \eqn{k=400}, but the user may wish to adjust this value
to be lower if the number of iterations in decreased, or higher if the number of iterations is
increased.
}
