% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_density_mode.R
\name{get_density_mode}
\alias{get_density_mode}
\title{Convenience function for estimating the mode of dataset's density using kernel estimation.
The kernel is estimated using R's base density() function and the max density point
is chosen. If multiple points have equally high estimated density, the central one is chosen.}
\usage{
get_density_mode(v)
}
\arguments{
\item{v}{Vector of distributed data as input for kernel density estimation}
}
\value{
A value specifying the position of the density mode
}
\description{
Convenience function for estimating the mode of dataset's density using kernel estimation.
The kernel is estimated using R's base density() function and the max density point
is chosen. If multiple points have equally high estimated density, the central one is chosen.
}
