% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exposure_pairwise_correlations.R
\name{get_exposure_pairwise_correlations}
\alias{get_exposure_pairwise_correlations}
\title{Pairwise correlation coefficients between signatures}
\usage{
get_exposure_pairwise_correlations(exposures_mcmc_output)
}
\arguments{
\item{exposures_mcmc_output}{Output from get_exposures()}
}
\value{
A dataframe of pairwise correlation values. One row per signature pair.
}
\description{
As a measure of signature bleed, this method computes the Spearman correlation coefficient
between each pair of mutation signatures across the MCMC sample space.
}
