% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exposure_summary_table.R
\name{get_exposure_summary_table}
\alias{get_exposure_summary_table}
\title{Creates a summary table of exposure statistics}
\usage{
get_exposure_summary_table(exposures_mcmc_output, alpha = c(0, 0.05, 0.5),
  fraction = FALSE)
}
\arguments{
\item{exposures_mcmc_output}{Output from get_exposures()}

\item{alpha}{Vector defining confidence bands to report. Alpha is between 0
and 1. The two confidence limits reported are
at rank alpha/2 and 1 - (alpha/2). For example, 
setting alpha = 0.05 results in 95% an interval capturing
the middle 95% of the data. Alpha = 0.5 results in
reporting of the 1st and 3rd quartiles. Multiple
alpha values can be provided as vector elements. These
values will each be computed and reported in separate columns.}
}
\value{
A dataframe with one row per signature and summary stats in the columns
}
\description{
After running get_exposures.R, one receives an object containing the points sampled
by the MCMC chains. It is often necessary to review summary statistics on these
sample chains to examine the exposure ranges of individual signatures.
This function provides an interface for computing summary statistics including
mean, median, range, and "confidence bands".
}
\details{
By default, the function reports mean and median, as well as range, 95% confidence bands,
and quartiles.
}
