% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exposures.R
\name{get_exposures}
\alias{get_exposures}
\title{Compute the MCMC sample space of exposure vectors}
\usage{
get_exposures(mutation_catalog = NULL, file = NULL,
  reference_signatures = NULL, n_chains = 4, n_iter = 10000,
  n_adapt = 10000)
}
\arguments{
\item{mutation_catalog}{Data frame with columns mutation_type, count. The mutation types
must be equivalent to those in reference_signatures.}

\item{file}{The mutation catalog can be stored as a TSV file, and its path can
can be provided in this parameter. Either mutation_catalog or file MUST
be provided.}

\item{reference_signatures}{Reference mutation signatures, such as that output from get_reference_signatures.}

\item{n_chains}{Number of MCMC chains}

\item{n_iter}{Number of iterations per chain. Total iterations will be n_chains times n_iter.}

\item{n_adapt}{Number of burn-in iterations}
}
\value{
List containing the MCMC samples, as well as other data such as reference signatures and mutation catalog.
}
\description{
At the heart of SignIT is the MCMC sampling of signature exposure solutions.
This function provides a convenient wrapper around the MCMC sampling steps.
It constructs the JAGS model, runs it, and then returns the result in a well formatted list.
}
