% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_populations.R
\name{get_population_parameter_estimates}
\alias{get_population_parameter_estimates}
\title{Point estimates for population Stan model output}
\usage{
get_population_parameter_estimates(population_mcmc_parsed)
}
\arguments{
\item{population_mcmc_parsed}{Output from \code{\link{get_populations}} after
processed by \code{\link{parse_stan_output}}.}
}
\value{
List of parameter estimates.
}
\description{
Parses Stan output and retrieves the mean for each parameter
}
