% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population_signatures.R
\name{get_population_signatures}
\alias{get_population_signatures}
\title{SignIT-Pop inference of populations and signatures}
\usage{
get_population_signatures(mutation_table, reference_signatures = NULL,
  subset_signatures = TRUE, n_populations = NULL,
  genome = BSgenome.Hsapiens.UCSC.hg19, method = "vb", n_chains = 10,
  n_cores = n_chains, n_iter = 300, n_adapt = 200)
}
\arguments{
\item{mutation_table}{Table of mutations, one per row. The minimum input requires the 
following columns: \itemize{
    \item \strong{total_depth}: Total number of reads covering mutated locus.
    \item \strong{alt_depth}: Total number of mutant reads covering locus.
    \item \strong{tumour_copy}: Tumour copy number at the mutated locus
    \item \strong{normal_copy}: Normal copy number at the mutated locus
    \item \strong{tumour_content}: Estimated tumour content as a fraction
                                   between 0 and 1. Must be the same value
                                   throughout the whole table.
}}

\item{reference_signatures}{Reference mutation signatures. This can either be from 
\code{\link{get_reference_signatures}} or a custom data frame formatted
equivalently.}

\item{subset_signatures}{Boolean. If TRUE (default), then \code{\link{subset_reference_signatures}}
is run to pre-select a smaller subset of signatures most likely to be active
in the cancer. This helps to reduce processing time and model complexity, but
may bias the result.}

\item{n_populations}{The number of populations to screen for. Must be an integer. If no value is
provided, then a model selection step is engaged to automatically estimate
the number of populations. The automatic model selection uses
\code{\link{select_n_populations}}, which performs a maximum a posteriori
estimate using the SignIT population model (without mutation signature inference).}

\item{genome}{A BSgenome object. This is used to determine trinucleotide contexts of mutations
to define mutation types. By default, uses BSgenome.Hsapiens.UCSC.hg19. To
define custom mutation types, simply include a column named \code{mutation_type} in
\code{mutation_table}, in which case this parameter is ignored.}

\item{method}{The posterior sampling method. This is a string and can either be 'vb' for
automatic variational Bayes or 'mcmc' for Hamiltonial Monte Carlo.}

\item{n_chains}{Number of chains to sample. Only relevant if \code{method == 'mcmc'}.}

\item{n_cores}{Number of cores for parallel sampling. By default this equals the number of chains.
Only relevant if \code{method == 'mcmc'}.}

\item{n_iter}{Number of sampling iterations per chain. These are distinct from adaptation iterations,
so the total number of iterations will be \code{n_iter + n_adapt}.
Only relevant if \code{method == 'mcmc'}.}

\item{n_adapt}{Number of adaptation iterations per chain. Only relevant if \code{method == 'mcmc'}.}
}
\value{
A list object with the posterior sampling of population signatures plus relevant input and metadata.
}
\description{
Jointly infers mutational subpopulations and their associated mutation signature exposures
}
\details{
\code{get_population_signatures} is the central function which facilitates Bayesian inference
of mutational populations and signatures. This model infers a matrix of L x N parameters,
where L is the number of populations and N is the number of signatures. The posterior distribution
of each parameter is estimated using either automatic differentiation variational inference
or Hamiltonial Monte Carlo using the \code{\link[rstan]{vb}} and \code{\link[rstan]{sampling}}
methods respectively of the \pkg{rstan} package (an interface to the Stan probabilistic
programming language).
}
