% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_populations.R
\name{get_populations}
\alias{get_populations}
\title{SignIT population model without mutation signatures}
\usage{
get_populations(mutation_table, n_populations, method = "mcmc",
  n_chains = 4, n_iter = 300, n_adapt = 200, n_cores = n_chains)
}
\arguments{
\item{mutation_table}{Table of mutations, same as input to 
\code{\link{get_population_signatures}}.}

\item{n_populations}{Number of populations to fit.}

\item{method}{The method to use. Can be either "mcmc" for Hamiltonial 
Markov Chain Monte Carlo or "MAP" for maximum a posteriori estimate.}

\item{n_chains}{Number of chains to use (only for \code{method == "mcmc"})}

\item{n_iter}{Number of sampling iterations per chain (only for \code{method == "mcmc"})}

\item{n_adapt}{Number of adaptation iterations per chain (only for \code{method == "mcmc"})}

\item{n_cores}{Number of cores for parallel sampling (by default equal to \code{n_chains}.
only for \code{method == "mcmc"})}
}
\value{
List object with Stan output, parsed data, and relevant metadata
}
\description{
Inference of mutational subpopulations without joint inference of signatures
}
\details{
This method can be used to obtain full posterior estimates of mutational subpopulations.
It is also used under the hood by \code{\link{select_n_populations}} to perform automatic
estimation of the number of populations for model selection.
}
