% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_signatures.R
\name{get_reference_signatures}
\alias{get_reference_signatures}
\title{Retrieves built-in reference mutation signature sets}
\usage{
get_reference_signatures(signature_set = "cosmic_30")
}
\arguments{
\item{signature_set}{The name of the reference signature dataset.
currently available options: 
cosmic_30 (default) - the WTSI 30 SNV signatures
cosmic_30_exome - WTSI signatures adjusted for exome trinucleotide frequencies
nikzainal_sv - Structural variant signature frequencies}
}
\value{
Data frame containing reference signatures in wide format, with a mutation_type column
}
\description{
Built in mutation signature sets provide reference data for
the analysis of N-of-1 mutation catalogs. This function provides
an interface for loading these reference datasets.
}
