% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rss.R
\name{get_rss}
\alias{get_rss}
\title{Computes the residual sum of squares MCMC solutions}
\usage{
get_rss(exposures_mcmc_output)
}
\arguments{
\item{exposures_mcmc_output}{Output from get_exposures}
}
\value{
A vector of RSS values, one per MCMC iteration
}
\description{
Iterates through every MCMC exposure solution and computes the
residual sum of squares between that solution and the mutation catalog.
}
