% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signature_bleed_graph.R
\name{get_signature_bleed_graph}
\alias{get_signature_bleed_graph}
\title{Graph Representing Signature Bleed Relationships}
\usage{
get_signature_bleed_graph(exposure_output = NULL,
  pairwise_correlations = NULL, min_bleed = 0.2)
}
\arguments{
\item{exposure_output}{Output from get_exposures}

\item{min_bleed}{Minimum bleed threshold for an interaction to be included in the graph}
}
\value{
An iGraph object where vertices are signatures and edges are bleed interactions.
}
\description{
Creates an iGraph object representing bleed between mutation signatures.
}
