% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mutation_types.R
\name{get_snv_mutation_type}
\alias{get_snv_mutation_type}
\title{Given chromosome, position, reference, and altered allele values, return a vector of
SNV mutation types conforming to the standard 96-element mutation types.}
\usage{
get_snv_mutation_type(chr, pos, ref, alt, genome = NULL)
}
\arguments{
\item{chr}{Character vector of chromosomes (must match genome seqlevels)}

\item{pos}{Integer vector of mutation positions}

\item{ref}{Character vector of reference alleles (A, C, G, or T)}

\item{alt}{Character vector of mutated alleles (A, C, G, or T)}

\item{genome}{BSgenome object from which to derive trinucleotide context sequences}
}
\value{
Character vector of mutation types
}
\description{
Given chromosome, position, reference, and altered allele values, return a vector of
SNV mutation types conforming to the standard 96-element mutation types.
}
