% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vaf_correction.R
\name{get_vaf_correction}
\alias{get_vaf_correction}
\title{Computes the VAF correction factor for a table of mutations}
\usage{
get_vaf_correction(mutation_table)
}
\arguments{
\item{mutation_table}{The same mutation table used as input to
\code{\link{get_population_signatures}}.}
}
\value{
Numeric vector of correction factors
}
\description{
The VAF correction factor adjusts the expected mean variant allele fraction
}
\details{
The correction factor is computed as

\deqn{\frac{p}{C^{(t)} p + C^{(n)} (1-p)}},

where \emph{p} is the tumour content / purity, $C^{(t)}$ is the tumour
copy number, and $C^{(n)}$ is the normal copy number.
}
