% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_population_signatures.R
\name{map_phi_to_signature}
\alias{map_phi_to_signature}
\title{Signature Index Mapper}
\usage{
map_phi_to_signature(phi_index, signature_names)
}
\arguments{
\item{phi_index}{Parameter index in output from \code{\link{get_population_signatures}}.}

\item{signature_names}{Vector of mutation signature names in order of appearance in reference signatures.}
}
\value{
Vector of signature names corresponding to parameter indices.
}
\description{
Maps parameter values to their respective mutation signatures
}
