% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_population_signatures.R
\name{map_population_signatures}
\alias{map_population_signatures}
\title{Maps Population Signature Parameters to their Signatures and Population Names}
\usage{
map_population_signatures(joint_model_output)
}
\arguments{
\item{joint_model_output}{Output from \code{\link{get_population_signatures}}.}
}
\value{
Same as input, with two new columns for signature name and population name.
}
\description{
Population signature model indices each correspond with specific signature and population indices.
}
