% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_exposures_as_matrix.R
\name{mcmc_exposures_as_matrix}
\alias{mcmc_exposures_as_matrix}
\title{Convert MCMC sampled exposure values to matrix format}
\usage{
mcmc_exposures_as_matrix(exposures_mcmc_output)
}
\arguments{
\item{exposures_mcmc_output}{Output from get_exposures()}
}
\value{
Matrix of MCMC iterations. One row per iteration, one column per signature.
}
\description{
MCMC samples of exposure solutions are initially in long format, with one row per
signature per iteration. This method converts this data frame to wide format,
with one row per iteration and one column per signature. It them converts this
to a matrix and returns it.
}
