% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_exposures_with_clustering.R
\name{merge_exposures_with_clustering}
\alias{merge_exposures_with_clustering}
\title{Convenience function which merges MCMC exposure iterations with density clusters}
\usage{
merge_exposures_with_clustering(exposures_mcmc_output, exposure_mcmc_clusters)
}
\arguments{
\item{exposures_mcmc_output}{Output from get_exposures.}

\item{exposure_mcmc_clusters}{Output from density_clustering.}
}
\value{
A tibble of MCMC iterations in wide format (one column per signature) and a column indicating assigned cluster.
}
\description{
Convenience function which merges MCMC exposure iterations with density clusters
}
