% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_types_match.R
\name{mutation_types_match}
\alias{mutation_types_match}
\title{Check whether mutation types match}
\usage{
mutation_types_match(catalogs_1, catalogs_2, order_is_important = TRUE)
}
\arguments{
\item{catalogs_1}{Mutation catalog object (data frame with cols: mutation_type, count)}

\item{catalogs_2}{Mutation catalog object (data frame with cols: mutation_type, count)}

\item{order_is_important}{Boolean. If FALSE, then the catalogs are sorted by mutation
type before they are compared.}
}
\value{
Boolean value. TRUE if mutation types match. FALSE otherwise.
}
\description{
Given two mutation catalogs, check that their mutation types match.
This is an important safeguard against potential errors which can arise
when catalogs have mismatched mutation types.
}
