% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_reference_signatures.R
\name{normalize_reference_signatures}
\alias{normalize_reference_signatures}
\title{Normalization of reference signatures.}
\usage{
normalize_reference_signatures(reference_signature_df)
}
\arguments{
\item{reference_signature_df}{Reference signature data frame, such as the
output of get_reference_signatures()}
}
\value{
Reference signatures in the same format, but normalized and non-zero
}
\description{
Ensures that all columns of reference signature matrix add to 1.
}
\details{
Also replaces all non-zero values are replaced with a very small number
(.Machine$double.eps).
}
