% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_population_signatures.R
\name{parse_stan_output}
\alias{parse_stan_output}
\title{Stan Output Parser}
\usage{
parse_stan_output(population_mcmc_out)
}
\arguments{
\item{population_mcmc_output}{Output from \code{\link{get_population_signatures}}.}
}
\value{
Tibble with chain information.
}
\description{
Converts Stan output to a long form tidy tibble
}
