% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_solution_errors.R
\name{plot_cluster_solution_errors}
\alias{plot_cluster_solution_errors}
\title{Summary plot of clusters}
\usage{
plot_cluster_solution_errors(exposures_mcmc_output, exposure_mcmc_clusters)
}
\arguments{
\item{exposures_mcmc_output}{Output from plot_exposures}

\item{exposure_mcmc_clusters}{Output from density_clustering}

\item{method}{Either 'bestfit' or 'mostdense'}
}
\value{
ggplot object. Scatterplot of clusters showing sum of squared errors vs. cluster score.
}
\description{
Summary plot of clusters
}
