% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_solutions.R
\name{plot_cluster_solutions}
\alias{plot_cluster_solutions}
\title{Plots solutions derived from density clustering}
\usage{
plot_cluster_solutions(exposures_mcmc_output, exposure_mcmc_clusters,
  method = "bestfit", k_value = 400)
}
\arguments{
\item{exposures_mcmc_output}{Output from get_exposures}

\item{exposure_mcmc_clusters}{Output from density_clustering}

\item{method}{Either 'bestfit' or 'mostdense'}
}
\description{
This plots the output of get_cluster_summary_points. See get_cluster_summary_points.R for
more details on the rationale behind this plot.
}
