% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_tsne.R
\name{plot_density_tsne}
\alias{plot_density_tsne}
\title{t-SNE plot of MCMC density clusters}
\usage{
plot_density_tsne(exposures_mcmc_output, patient_density_clustered,
  perplexity = 30)
}
\arguments{
\item{exposures_mcmc_output}{Output from get_exposures}

\item{patient_density_clustering}{Output from density_clustering}
}
\value{
A ggplot object containing the t-SNE plot in 2 dimensions
}
\description{
t-SNE is a useful tool for visualizing data clusters in high dimensional space.
This method produces a t-SNE plot showing the relationship of points sampled
by MCMC and coloured by density clusters.
}
