% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exposure_chain.R
\name{plot_exposure_chain}
\alias{plot_exposure_chain}
\title{Plot the MCMC chains for the signature exposure solutions}
\usage{
plot_exposure_chain(exposures_mcmc_output, signature_trim = "Signature")
}
\arguments{
\item{exposures_mcmc_output}{Output from running plot_exposures}
}
\value{
A ggplot object
}
\description{
This function allows one to visualize the solution chains for the exposure
of each mutation signature. It plots each chain as a line graph, allowing
one to inspect for convergence.
}
