% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exposure_posteriors.R
\name{plot_exposure_posteriors}
\alias{plot_exposure_posteriors}
\title{Plot the signature exposure posterior distributions}
\usage{
plot_exposure_posteriors(
  exposures_mcmc_output,
  view = "violin",
  signature_trim = "Signature",
  units = "mutations"
)
}
\arguments{
\item{exposures_mcmc_output}{Output from get_exposures.}

\item{view}{Can be either "violin" or "boxplot".}

\item{units}{Units to present exposures in. Can be
'mutations', 'megabase', or 'fraction'.}
}
\value{
A ggplot plot of the posterior distributions.
}
\description{
Visualizes the posterior distributions of the MCMC solution to signature exposures.
By default, they are visualized as violin plots.
}
