% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_population_signatures.R
\name{plot_population_signatures}
\alias{plot_population_signatures}
\title{Plot Population Signatures}
\usage{
plot_population_signatures(joint_model_output)
}
\arguments{
\item{joint_model_output}{Output from \code{\link{get_population_signatures}}.}
}
\value{
ggplot object with population mutation prevalences, proportions, and signature exposures.
}
\description{
Creates a violin plot of the full posterior estimates for population signature exposures.
}
