% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signature_pairwise_bleed.R
\name{plot_signature_pairwise_bleed}
\alias{plot_signature_pairwise_bleed}
\title{Plots pairwise correlations between signatures in MCMC samples}
\usage{
plot_signature_pairwise_bleed(exposures_mcmc_output)
}
\arguments{
\item{exposures_mcmc_output}{Output from get_exposures.}
}
\value{
A ggplot object showing a heatmap of pairwise correlation coefficients.
}
\description{
This plot allows the user to explore their MCMC solutions for the presence
of mutation signature bleed effects.
}
\details{
The MCMC sample space is likely to contain some degree of multicollinearity.
That is to say that some signatures are likely to correlate or anti-correlate
with others. This can result from a phenomenon known as "Signature Bleed,"
wherein signatures may feed into one anothers' signals if they share similar
mutational spectra.

This function computes correlation metrics between each pair of signatures.
It then plots them as a grid. Signature bleed is associated with anticorrelation
between a signature pair, which suggests that good solutions can substitute
the contribution of one signature for the other.
}
