% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_two_signature_hexplot.R
\name{plot_two_signature_hexplot}
\alias{plot_two_signature_hexplot}
\title{Plots a hexplot between two signatures}
\usage{
plot_two_signature_hexplot(exposures_mcmc_output, signature_name_1,
  signature_name_2, exposure_mcmc_clusters = NULL, cluster_number = NULL,
  trendline = TRUE)
}
\arguments{
\item{exposures_mcmc_output}{Output from get_exposures.}

\item{signature_name_1}{String denoting signature name (must match one of those in exposures_mcmc_output)}

\item{signature_name_2}{String denoting signature name (must match one of those in exposures_mcmc_output)}

\item{exposure_mcmc_clusters}{Optionally, you can provide the output from density_clustering, which allows
you to visualize the comparison plot for a single cluster.}

\item{cluster_number}{If you wish to visualize a single cluster, provide the cluster number here.}

\item{trendline}{Boolean value specifying whether to plot a linear trendline.}
}
\value{
A ggplot object showing the correlation of MCMC samples between two signatures.
}
\description{
The MCMC sample space allows one to examine for multicollinearity between
mutation signatures. For more details on signature bleed, see plot_signature_pairwise_bleed.
}
\details{
This function provides a closer view on correlated signatures in the
MCMC sample space. It plots the relationship between two signatures as a
hexplot, and reports the Spearman correlation coefficient.
}
