% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_signatures_as_matrix.R
\name{reference_signatures_as_matrix}
\alias{reference_signatures_as_matrix}
\title{Convert reference signatures to matrix}
\usage{
reference_signatures_as_matrix(reference_signatures, mutation_catalog)
}
\arguments{
\item{reference_signatures}{Reference mutation signature as output by get_reference_signatures()}

\item{mutation_catalog}{A mutation catalog with mutation types matching those of the reference
signatures. This catalog is used only to order the rows of the output
matrix such that they correspond with the mutation type order in mutation_catalog.}
}
\value{
A matrix with named columns, one for each signature.
}
\description{
Converts a data frame of reference signature data to the matrix form.
}
\details{
When generating the matrix, a mutation catalog is required. Because the matrix output
does not have a mutation_type column, the matrix rows are arranged so that mutation types
match the order of the provide mutation catalog. This way, the matrix output can immediately
in correspondence with the provided mutation catalog.
}
