% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_population_signatures.R
\name{reverse_factor_levels}
\alias{reverse_factor_levels}
\title{Reverse Factor Levels}
\usage{
reverse_factor_levels(x)
}
\arguments{
\item{x}{A factor}
}
\value{
A factor with levels reversed
}
\description{
Takes a factor, reorders the levels, and reassigns values to the new levels.
}
