% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_population_signatures.R
\name{reverse_factor_ranks}
\alias{reverse_factor_ranks}
\title{Reverse Factor Ranks}
\usage{
reverse_factor_ranks(x)
}
\arguments{
\item{x}{A factor}
}
\value{
A factor with levels re-ranked
}
\description{
Takes a factor and re-ranks its levels. None of the values change, but the level ranks are reordered.
}
