% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_x_axis_labels.R
\name{rotate_x_axis_labels}
\alias{rotate_x_axis_labels}
\title{Rotates the x axis labels 90 degrees in ggplot graphs.}
\usage{
rotate_x_axis_labels()
}
\value{
ggplot component which can be added to a ggplot graph with the "+" operator.
}
\description{
Convenience function for plots.
}
