% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_signature_model.R
\name{run_signature_model}
\alias{run_signature_model}
\title{Run the JAGS model to decipher mutation signatures by MCMC}
\usage{
run_signature_model(signature_model, signature_names, n_mutations,
  n_iter = 10000)
}
\arguments{
\item{signature_model}{The JAGS model obtained from get_signature_model}

\item{signature_names}{Signature names provided as a character vector. Must match
the number of signatures provided to the model.}

\item{n_mutations}{The number of somatic mutations. This is multiplied by the
fractional exposure vector to obtain the absolute number of mutations
associated with each mutation type.}

\item{n_iter}{The number of iterations to run per chain.}
}
\value{
Data frame with one row per signature per iteration of the chain.
}
\description{
Run the JAGS model to decipher mutation signatures by MCMC
}
