% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_n_populations.R
\name{select_n_populations}
\alias{select_n_populations}
\title{Automatic Model Selection for Number of Populations}
\usage{
select_n_populations(mutation_table, parallel = TRUE)
}
\arguments{
\item{mutation_table}{Input tibble, same as for \code{\link{get_population_signatures}}.}

\item{parallel}{Set TRUE to run in parallel on the available cores.}
}
\value{
List object reporting the optimal model, evaluated model outputs, and the BIC of each
}
\description{
Attempts to estimate the number of mutation subpopulations using Bayesian information criterion.
}
