% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exposures.R
\name{signature_present}
\alias{signature_present}
\title{Determine Presence or Absence of Signatures}
\usage{
signature_present(exposure_chain, z_value = 1.96)
}
\arguments{
\item{exposure_chain}{Exposure chain table, part of output from get_exposures().}
}
\value{
Table indicating the presence or absence of each signature.
}
\description{
Determine Presence or Absence of Signatures
}
