% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signit-package.R
\docType{package}
\name{signit-package}
\alias{signit-package}
\alias{signit}
\title{SignIT: Mutation Signatures in Individual Tumours}
\description{
A package for robust and intepretable estimation of mutation signatures and their temporal evolution.
}
