% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mutation_types.R
\name{stitch_mutation_types}
\alias{stitch_mutation_types}
\title{Given a set of substitutions and trinucleotides, merge into unique mutation
types. For example, T>A mutation in CTG context will become C[T>A]G.}
\usage{
stitch_mutation_types(substitution, trinucleotide)
}
\arguments{
\item{substitution}{Character vector of substitutions in the form T>A}

\item{trinucleotide}{Character vector of trinucleotide contexts in the form CTG}
}
\value{
Character vector of mutation types in the form C[T>A]G
}
\description{
Given a set of substitutions and trinucleotides, merge into unique mutation
types. For example, T>A mutation in CTG context will become C[T>A]G.
}
