% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_reference_signatures.R
\name{subset_reference_signatures}
\alias{subset_reference_signatures}
\title{Automatic subsetting of reference mutation signatures}
\usage{
subset_reference_signatures(catalog, reference_signatures, threshold = NULL)
}
\arguments{
\item{catalog}{Mutation catalog. A data-frame or tibble with columns
named mutation_type and count.}

\item{reference_signatures}{Reference mutation signatures, formatted as in those
obtained by \code{\link{get_reference_signatures}}.}

\item{threshold}{A numeric value specifying the minimum lower credible interval
(at the 95% confidence level). Any signatures with a 2.5%
interval limit above this threshold will be included in the subset.}
}
\value{
A new reference signature dataframe with a subset of the initial signatures
}
\description{
Subsets mutation signatures based on which ones are present in a somatic mutation catalog
}
