% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_population_signatures.R
\name{summarise_population_signatures}
\alias{summarise_population_signatures}
\title{Summarise Population Signatures}
\usage{
summarise_population_signatures(joint_model_output)
}
\arguments{
\item{joint_model_output}{Output from \code{\link{get_population_signatures}}.}
}
\value{
Tibble of summary statistics
}
\description{
Report summary statistics (mean, median, mode, and SD) for population signature parameters
}
