% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_signature_names.R
\name{trim_signature_names}
\alias{trim_signature_names}
\title{Trims the names of mutation signatures}
\usage{
trim_signature_names(signature)
}
\arguments{
\item{signature}{A factor containing signature names.}
}
\value{
A factor where all names starting with "Signature" are trimmed.
}
\description{
Any signature names starting with "Signature" get trimmed.
}
