% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exposures.R
\name{get_exposures}
\alias{get_exposures}
\title{Compute the MCMC sample space of exposure vectors}
\usage{
get_exposures(mutation_catalog, reference_signatures = NULL, priors = NULL,
  n_chains = 4, n_iter = 200, n_adapt = 200, n_cores = 1,
  stan_model = NULL, quiet = FALSE, z_value = 1.96)
}
\arguments{
\item{mutation_catalog}{Data frame with columns mutation_type, count. The mutation types
must be equivalent to those in reference_signatures.}

\item{reference_signatures}{Reference mutation signatures, such as that output from get_reference_signatures.}

\item{n_chains}{Number of MCMC chains}

\item{n_iter}{Number of iterations per chain. Total iterations will be n_chains times n_iter.}

\item{n_adapt}{Number of burn-in iterations}

\item{z_value}{The z value for determining presence/absence of signatures.
Determines the confidence bounds (z times standard deviation).
Signature is called absent if zero falls within the interval.
Higher z_values increase specificity and lower z_values increase sensitivity.}
}
\value{
List containing the MCMC samples, as well as other data such as reference signatures and mutation catalog.
}
\description{
At the heart of SignIT is the MCMC sampling of signature exposure solutions.
This function provides a convenient wrapper around the MCMC sampling steps.
It constructs the STAN model, runs it, and then returns the result in a well formatted list.
}
