% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signature_model.R
\name{get_signature_model}
\alias{get_signature_model}
\title{Get the STAN Model for Mutation Signature Analysis}
\usage{
get_signature_model()
}
\value{
STAN model object
}
\description{
Creates the STAN model for MCMC analysis of mutation signatures.
}
