% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signature_names.R
\name{get_signature_names}
\alias{get_signature_names}
\title{Parse Signature Names from Reference Signatures}
\usage{
get_signature_names(reference_signatures)
}
\arguments{
\item{reference_signatures}{Reference mutation signatures (i.e. from get_reference_signatures)}
}
\value{
Character vector of mutation signature names
}
\description{
Retrieves the names of mutation signatures from reference signatures
}
